Add Examples to Ref. Guide for TRAN_CONT Measures
This commit adds two examples to the Reference Guide for TRAN_CONT
measure mode.  The first one illustrates the output file names for
TRAN_CONT measure mode.  The second example illustrates the
interpretation of negative and non-negative values of the CROSS
qualifier for both TRAN and TRAN_CONT measure modes.

SON Bugs 1274 and 1313
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313