Another fix related to issue #206
Issue #206
"Parser gets confused with node names that match model names (at least
in one example)"

Issue # 144
"Make model binning capability turned on by default"

My fix for issue #144 a while ago involved re-arrangement of the
DeviceBlock::extractModelName function.    My fix for #206 involved
fixing a logic error in that function, in which the function would give
up a bit too early on finding a model name.

The various efforts have resulted in the DeviceBlock::extractModelName
function having the following structure:

1. check the device instance line fields to see if any are a valid model

2. If not, and the device is a MOSFET, then attempt to do model binning

3. If both (1) and (2)  fail, then the function fails.

4. If either (1) or (2) succeeds, the function succeeds.

My recent work meant that if a MOSFET instance failed to find a model in
step (1), it would always proceed to step 2.    Apparently before, it
didn't always do this.  This is the right thing to do, but it revealed a
downstream problem.

It turns out that the model binning code had a vulnerability, which
was that a std::string substr function was unconditionally invoked.
It should only have been invoked if the arguments were valid.
Otherwise it throws an exception, and I had not setup the code to
catch that exception.  The arguments were likely to be invalid if
it failed to find a binned model.

So, this commit makes a very small change, to put an if-statement around
the call to substr, to make sure the arguments to the function are
valid.  If they are not valid, the function fails.  If the function
fails, then the code exits with an error.  But, unlike before, it is
a graceful error with a coherent error message, rather than a
mysterious exception.