Restore Xyce 7.3 -per-processor Behavior for .MEASURE
This commit fixes a parallel debugging nit that is mostly invisible
to the end-users.  Invoking Xyce in parallel with a command line like
this:

  mpirun -np 2 /path/to/Xyce -per-processor -l logfile <netlistName>

will now result in all of the measure results being printed to the
logfile for processor 2.  This includes the results for both continuous
mode and non-continuous mode measures.  This restores the logging
behavior to what is was before .OPTIONS MEASURE USE_CONT_FILES=<value>
was added to the develop branch with commit 34faa261 this week.

SON Bug 1274
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274