Remove obsolete LaTeX comments
Back in the day, we started the practice of putting information about
the document's R&A status in LaTeX comments, just to keep track of
that information somewhere.  This information included submission
date, tracking number, and approval date.

That information is not actually as important as it used to be now
that R&A of documents that fall under the DUSA takes about a day.  And
all of that information can be tracked more easily in a single place
on the Gitlab issue tracker item for the release checklist.

I'm removing the comments placeholder for that information to
emphasize that it is no longer needed here.

That information should still be kept track of, but the place to keep
track of it is the issue tracker.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/198