Fix vcrit expression in Verilog-A diode examples
Years ago, I added voltage limiting code to the two example Verilog-A
diode models in the "toys" directory.  The expression for vcrit has
always been incorrect.

Over two years ago, this error was pointed out to us by an open
source user on the google group, but I never went back and fixed it.

It turns out not to matter very much in this example, which is only
used in one demo netlist where limiting doesn't even matter, but it is
wrong nevertheless and should have been fixed at the time it was
pointed out.  Better late than never.

https://groups.google.com/g/xyce-users/c/GksAM0aPcC8?pli=1
https://github.com/Xyce/Xyce/discussions/25