Fix for issue #213
Issue #213
"delete the OutputMgr data structures mainContextFunctionMap_,
mainContextParamMap_, and mainContextGlobalParamMap_ after setup phase"

This commit adds a few things.

1.  It adds new function calls to the output manager, which clear the 3
    maps, which are the subject of issue #213.

2. These function calls happen at the end of the "initializeLate"
   function.

3. That means that all use of these maps need to be over by the end of
   this initializeLate function.

4. It turns out that this wasn't quite the case.  The way that the
   outputters are managed meant that they are set up later in the
   calculation, kind of "on the fly".    So, this commit includes a lot
   of refactoring of the .PRINT outputter setup.

Here are some of the details:

This commit cleans up the outputter setup, and also creates a new set of
functions in the analysis classes for handling the "last chance"
expression setup, before all the Maps are deleted.  These functions need
to be called before any of the actual analysis happens, from the
"initializeLate" function in N_CIR_Xyce.C.

Basically all of the work on this bug has been to make sure that any expressions
that depend on .global_param, .param or .func are 100% set up and resolved by
the end of the "initializeLate" function.   That way, the 3 maps, which are
the subject of this bug, can be deleted at the end of that function, and Xyce
can use a lot less memory for some circuits.  (particularly large PDK circuits).

A lot of "last chance" processing of expressions is for .PRINT outputters.
This commit modifies every analysis which uses the "Active Outputter" logic
(in other words, all of them) to now have an additional function call
early during setup, to make sure all the outputters that might be needed
are allocated, before the maps are deleted.

This commit also takes care of a similar issue for .RESULT.   One weird
aspect of .RESULT is that the analysis manager handles its allocation as
a "processor" object.  AFAIK, the processor objects in the analysis
manager are only .RESULTS and nothing else.  But they use a fancy base
class, that the analysis objects ALSO use.   So at the most abstract
level, a .RESULT is simply another analysis on the stack.  I'm not sure
what the point of doing this was, but there it is.  Anyway, just like
the .PRINT outputters, the setup of .RESULT needed to happen earlier
as well, and that is done now.

With these final changes, the whole test suite passes again.