Cleanup/fix obj. function header for issue #213
Issue #213
"delete the OutputMgr data structures mainContextFunctionMap_,
mainContextParamMap_, and mainContextGlobalParamMap_ after setup phase"

This header contains functions that are used for  setting up sensitivity
analysis objective functions.  The most important functions are:

setupObjectiveFunctions
setupObjectiveFuncGIDs

and these are the functions modified by this commit.  The main thing
that gets fixed is that the setupObjectiveFuncGIDs has been fixed so
that it no longer double checks to see if any of the strings in the
"expVarNames"  std::vector<std::string> vector are global parameters.

This function was written to loop over this vector,
and then obtain a GID for each one. (GID = global solution vector index)

In the old code (before this commit), it would check each of these to
make sure it WASN'T a global parameter first.  If it is a global
parameter, then don't try to find the GID.  However, this was a silly
b/c the function that set up this std::vector already had this
information.

So, this function removes the "check" code from setupObjectiveFuncGIDs,
and it also removes the code from setupObjectiveFunctions that
(pointlessly) added global_parameters to the expVarNames object.