Remove expression eval from addGlobalParameter
Global parameters are added to the device manager via the
DeviceMgr::addGlobalPar function, which subsequently calls the
addGlobalParameter free function.

In the addGlobalParameter free function, the process includes checking
if this parameter is of type EXPR, and if it is, then adding the
expression to a container of expressions, and then evaluating the
expression prior to setting its value in a map.

The evaluation isn't necessary at this stage and is actually really
fragile.  I'm refactoring some parts of the parser, and my refactoring
exposed this fragility.  It is fragile b/c any expression can depend on
global parameters, but the container that the expression will rely upon
to resolve those parameters is the very container that this function is
modifying.  So, this expression evaluation only worked correctly if the
global params were added in the correct order.  With my recent
refactors, I broke the order.

But there was no reason to evaluate it here.  Expressions get evaluated
later anyway in functions like "updateDependenParams".

This is related to Issue #35
"Reduce netlist setup time for modern PDKs"

It is also kind of related to the closed Issue #167
"Make globally-scoped .params be equivalent to .global_param"