Update global params to use unordered_set
Issue #35
"Reduce netlist setup time for modern PDKs"

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

Around 9 months ago, Heidi updated the parser so that .params would be
stored in a std::unordered_set, rather than a std::list.  This was to
mitigate the cost of "find" based searches in large PDKs, which could
have 1000s of parameters.  This had a very significant impact on
runtimes for circuits using these PDKs.

Much more recently, I worked on Issue #167.  As an experiment for that
issue, I tried out having ALL .params in the top context be elevated to
.global_params.  When I did this, I ran into various problems, so I
abandoned that approach to the issue.  One of the problems was that when
I did this, circuits using really large modern PDKs went back to being
really, really slow to setup.  This puzzled me, but I didn't have time
to track down all the reasons for this.

It looks like one reason was that global_params were still being stored
in std::lists.  So, this commit updates the code to put global_params
into std::unordered_sets, just like .params.

This did uncover one unexpected issue, which I addressed in my previous
commit.  The process for adding global parameters to the device package
was implicitly order-dependent.  Any paramter added to the device
package had to be added *after* all of its dependent params were added.
Switching to a std::unordered_set broken this order.  But it was easy to
fix, once I knew what was going on.  It was fixed by removing a
superfluous expression evaluation, in my previous commit.