Enable a limited Xyce CMake build for Charon
Charon can utilize a certain Xyce capability, but it is inconvenient for the
project to build all the required Xyce TPLs and Tilinos packages. This commit
enables a build of Xyce that has just enough features to compile and support
the capability needed by Charon. It can be enabled with a flag in the CMake
invocation: `-D Xyce_AS_SPECIAL_CHARON_TPL=ON`. A warning will be giving during
the configure stage if it is on, since the Xyce build might be unusable for any
other purpose.

Since Charon builds Trilinos as a shared library, I also had to change the
first steps in the root CMakeLists.txt file. Basically, `project` has to be
called before `find_package(Trilinos...)` to ensure CMake knows enough about
the system that it will allow shared builds. Therefore, `project` is now called
twice, once at the beginning, with "NONE" for the compilers, and then the
regular call after the compilers have been defined by Trilinos.

In addition to the above, I also did a bit of cleanup and organization. In
particular, I made CMake not probe for certain capabilities if they have been
expressly turned off. The previous process did the right thing, but made it
seem like features were being turned on when they were not.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/216