Slight refactor newExpression::make_var function
Issue #35
"Reduce netlist setup time for modern PDKs"

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

I am attempting to figure out why, for a large PDK, .global_param is so
much less efficient than .params.

If I force the N_IO_CircuitContext class to automatically convert ALL
.params from a netlist into .global_params, small circuits work OK, but
large netlists get slower.  This is a bit of a mystery to me at the
moment that I'm attempting to track down.   This is really noticable for
the BUG_1222_SON/translated_vco/uw_VCO.cir regression test.  It runs
reasonably quickly with .param, and bogs down like crazy with
.global_param.  I am not sure why yet.

For this commit, I have updated the "make_var" function so that it
passes in a value to the parameter.  This renders "make_var" a little
more similar to "make_constant".  (although they are still different)
Previously, "make_var" only set the status of a parameter inside of an
expression and did not set a value.    "make_constant" would both set a
status and a value.  This kind of made sense; "make_constant" was
basically saying, "never update the value of this parameter", while
"make_var" basically says "expect to update the value for this
parameter".

But during parsing, it makes sense to do the best we can to
assign a value, even on a "variable" parameter.

Both the make_var and make_constant function calls are holdovers from
the previous expression library.  They are also one of the most noticable
differences between how .param and .global_param are handled in the parser.

For any expression, the parser will attempt to resolve any dependent
parameters.  For both global and normal params, if the parameter is
expression dependent, then the parser simply attaches the parameter and
moves on.  It does not attempt to evaluate the expression at this point,
although it may do some diagnostics (like to ensure it is not solution
dependent).

But if a parameter is deemed to NOT be expression-based, then they are
handled differently.  If the type is DBLE or STR, then .params receive
the "make_constant" treatment, and .global_params receive the "make_var"
treatment.

One problem with this is that if an expression depending on
a global then needs to be evaluated during parsing then it might
be impossible, because any "variable" parameter cannot correctly
retrieve information from Xyce (via the mainExpressionGroup object).  It
isn't wired to just use the DBLE value that was known at the time of the
make_var call.  So this leads to weird circular behavior, which I think
may be a problem.  The expression will tell the group object to retrieve
the value of the "variable" global_param, and the expression group will
then call the device package to get this value.

In some cases, this results in the device package being unable to
retrieve the value.  Sometimes this can lead to a fatal error. Other
times, it might mean that stuff gets allocated in the device package
that (arguably) shouldn't be allocated yet.  Either way, this is an
potentially expensive operation that doesn't need to happen.  Most of
the time, at least during parsing, the code should just use the best
values it knows at the time, whatever they are, and it should not need
to do a complicated (confusing) retrieval of these values.