Use getMutableParam in various places
This change is related to:

Issue #35
"Reduce netlist setup time for modern PDKs"

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

As part of this work, I've discovered that .param is allowed in many
more situations than .global_param.  In some cases, this was
intentional, and in others it is accidental.

This commit addresses various cases that were arguably intentional,
but should be changed anyway.