Fix CMake logic error in previous commit
I had a logic error in the previous commit. The CMake has been written
such that, if a Trilinos package is available (and usable by Xyce), then
it is automatically enabled. The logic error broke that approach. This
commit returns it to the desired state.

I also removed some extraneous CMake output.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/216