Add frequency mapping algorithm for HB
The artificial frequency mapping algorithm for box truncation method
is added in the commit. It maps the real frequencies in the HB analysis
to artifical frequencies and leads to uniformly spaced frequency
spectrum. For multi-tone HB analysis, the real frequency spectrum is
not uniformly spaced. Therefore, the FFT/IFFT cannot be used. With
the artificial frequency mapping algorithm, the FFT can now be applied
to the multi-tone HB analysis.

The current framework of HB analysis needs to be changed in order
to handle AFM based HB. The artificial frequencies are not related
to the real time by the FFT. So the current method of handling B vectors
would not work for AFM algorithm. The new algorithm needs new device
loading technique and supporting framework to work properly

Xyce#160