Fix a couple of push_back/insert issues
In commit 8f1328b1, @erkeite had changed a container to an
unordered_set, and doing so required changing some push_backs into
inserts.

When I just tried to build Xyce after the
release->develop merge, I got a compile failure because a *few*
push_backs still existed.

This is because a few new blocks of code had been added on the release
branch and needed the same treatment after the merge.