More refactoring of group classes
Related to these issues:

Issue #35
"Reduce netlist setup time for modern PDKs"

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

This commit includes a lot of code cleanup, as well as further
refinements on my recent "group" refactor.   Partly, I decided to rename
and re-arrange the groups a bit, to make their names and usage a little
less confusing.  So now, we have the following groups:

expressionGroup  <-- base class, unconditionally returns 0 for
everything.

paramParsingExpressionGroup <-- used only during parsing, in the circuit
context.  Very similar to expressionGroup in that it does almost
nothing.  This is reasonable at this stage, b/c it isn't possible to
resolve a lot of things during the parse stage.  The most important
thing that is different between the expressionGroup and this group is
that this group will return the given value for a global_param, not 0.0.

mainXyceExpressionGroup <-- used by things like sensitivity analysis,
and various other parts of the code.  This is also passed into the
device package, but mostly as a container with various useful references
in it, that can be used to create a device-oriented group.   This group
relies a lot on lookups to retrieve values, so it isn't very efficient.

deviceExpressionGroup <-- used by devices.  This is derived from the
mainExpressionGroup.  But unlike the main group, it uses LIDs and relies
less on lookups.  So, it is a bit more efficient.