Modify arguments for computeXyceVectorsWithLimiting(...)
solV provided to computeXyceVectorsWithLimiting used to be the
variable nextSolutionVariables, but now solV is a vector of vectors
with index [0] being next, [1] being current, and [2] being last.

Similarly, staV now provides stateVariables with the same indexing
pattern.

None of these changes affect computeXyceVectors(...) which continues
to use solV as a 1D array representing nextSolutionVariables.