Refactor of voltageOp class
This is kind of related to Issue #35
"Reduce netlist setup time for modern PDKs"

and other performance related issues.  But mostly this commit is about
making the code cleaner and simpler.

Early in development of this expression library, I started down one path
for handling voltage drops specified as V(A,B), where the two nodes are
inside the parens and separated by a comma.  Later, I realized this
wasn't the right way to go, and went down a different path.  But many
artifacts of the old code were still there.

The old idea (which I'm removing old artifacts of) was to handle the
voltage drop entirely inside the voltageOp class.  This meant that the
voltageOp class needed to own an array of voltage values, which might be
size=1 or size=2.  And, throughout the code, there would be
if-statements, etc to check the size.

My later idea was to NOT handle this difference inside of the voltageOp
class, and instead have voltage drops handled by the AST.  So, if the
parser encounters V(A,B), it then creates an AST where a subtraction
operator has V(A) and V(B) as its two arguments.  This makes things a
lot easier, especially when dealing with different types (double vs.
complex), and especially easier for dealing with derivatives.  Also
easier when dealing with things like VDB(A,B).

So, the code has been using this second approach for a long time at
this point, but the voltageOp class was still wired to having an array
of values.  But, much of that code was never used.  So it needed to go.
This commit removes it, and makes voltageOp (as well as code that
invokes it) a bit simpler and cleaner.

This commit also removes an obsolete unit test.