Remove the paramParsing group, merge w/ base class
This is kind of related to Issue #35
"Reduce netlist setup time for modern PDKs"

It is also related to various other tracked issues having to do with
expression library performance.

I've been working to cleanup and simplify the "Group" concept for the
new expression library.  In my initial work to integrate the new library
into Xyce, I was mostly focused on getting the entire test suite to
pass, and I left a lot of code cleanup, and potential efficiencies,
on the table.

In a previous set of commits, I created a new group that is used only
for the parse phase. I called it the paramParsingExpressiongroup.   It
is a minimal group, intended to make parsing more efficient.  The group
does almost no work, so any expression evaluations won't incur any
"group" related costs for looking up parameter values, etc.

Funcionally, it was very similar to the base group, except for a handful
of things that it did differently.  So it made sense to merge them and
make the code a little less complicated.

As of this commit, the following group classes exist:

baseExpressionGroup <- basically does nothing and is used in parsing.
It is used in parsing b/c it should be fast and b/c the circuit problem
is not set up yet, so many of the data retrievals supported by the other
groups don't make sense during parsing.

mainExpressionGroup <- default group used by things like sensitivity
objective functions.  It can work in most contexts, but is slow.

deviceExpressionGroup <- used by expressions that are applied to device
model parmaeters.  Optimized to be faster than the main group, and
derived from it.

ACExpressionGroup <- basically same as mainExpressionGroup, but set up
for frequency domain problems.

outputsExpressionGroup <- supports many more types of operators than the
other groups, and also only works on proc 0.