Include "gmin" in the list of "specials"
This is kind of related to Issue #35
"Reduce netlist setup time for modern PDKs"

In various places during parsing, when parameters are resolved, the
right hand side is parsed as an expression, and then queried to see if
it includes anything that would prevent it from being set as a simple
DBLE or STR type parameter.  If it can't be set to DBLE or STR, then it
is set to EXPR type.

Dependencies that would cause this (EXPR type) to happen have included
voltage nodes, device currents, and global parameters.  Also included in
this list are various "specials" such as TIME and FREQ.  Many of these,
of course, are not allowed in certain use cases.  But sometimes they
are.

One dependency that I had NOT been treating as a "special", and was thus
excluding from the above list was GMIN, which must be obtained from the
device package to resolve correctly.  As I am attempting to simplfy
expression handling by using a "do nothing" group during parsing, I
needed to include GMIN in the specials list.  Otherwise, it would simple
attempt to evaluate it during parsing, and use the value gmin=0.0, and
then never updated it (because it would be type DBLE rather than EXPR).

This commit fixes the issue.  Gmin is now treated as a special.