Add a framework to enable frequency domain loading of sources
The voltage and current sources are only evaluated in time domain
in all previous HB algorithms. However, this does not work with the
AFM algorithm. A different framework to evaluate sources are needed.
This commit adds a framework to evaluate the sources in frequency
domain to support the new algorithm. The Jacobian is evaluated in
time domain and the B vectors can be evaluated in frequency domain for
linear sources in HB with the framework

Xyce#160 Xyce#219