Fix for issue84_devParam test on four processors
I recently changed the parser so that global parameters are stored in a
std::unordered_set container.  This enabled faster searches, but created
a situation where the order in which these global parameters are added
to the device package was not consistent from one processor to another.

For most aspects of Xyce, this didn't matter, as most operations on
global_params in the device package are insensitive to order.  But
for some UQ procedures, this mattered.  It was possible to
have different sets of sample points (probably the same points, but in a
different order) being used on different processors,
which made the calculation inconsistent.

Anyway, the easiest way to fix this was to modify the function
DeviceMgr::getRandomParams.    In this function, a std::vector of
Xyce::Analysis::SweepParam objects is populated.  It can contain global
params, device model params, or device instance params.  For device
models and instances, I knew that handling this in parallel was tricky,
so this function already contained a lot of code for handling them in
parallel, including sorting.

However, for the global parameters, I had not bothered doing anything
special in parallel, including sorting.  This was because global params
are distributed to every processor, and prior to my change to the
parser, sorting wasn't necessary.   So, to fix this issue I just had to
add another std::sort to this function.  That is contained in this
commit.