Getting rid of make_var function
This is part of a larger refactor.  In the new expression library I
mostly tried to move away from the old API, but there were a few
features of that API that I kept, mostly because I ran out of time.

The "make_var" function was one of these features.  In general,
parameters are handled by the "paramOp" class, but I had to make it
operate in 3 different modes:

attached
constant
variable

"Attached" was my preferred method, and kind of the point of the new
library.  When parameters are resolved in an expression, they can just
attach the AST of the defined parameter.    Then the value doesn't need
to be "set".  It is just naturally part of the tree.

"constant" is used by .param, for parameters that evaluate to a
simple number where the type = Util::DBLE or Util::STR.  Then there is
no tree to attach, it just gets hardwired to that value.

"variable" has been used by .global_param, in much the same was as
"constant" except that there is an expectation that it will change, and
the change will happen outside the current expression.  So, these
parameters need to "pull" their value from Xyce usually via the
DeviceMgr::getParamAndReduce function.  This is a bit awkward, and
involves searches every time.   And, as these parameters could be
handled much more cleanly via the "attached" functionality, it always
bugged me that this function was still in place.

I'll note that the "variable" tag is has been ambiguous, in that it has
meant 2 things:  (1) this is a DBLE typed global param and (2) we want
derivatives w.r.t. this param.  This work disamgiuates the variable
boolean.  It now ONLY means (2).

This commit also renames the "getGlobalParameterValue" group functions to
"getParameterValue" as this is more accurate.
At this point ALL global parameters are handled by attachments, so these
functions aren't used to retrieve any global param values anymore.
However, they are used to retrieve other types of parameters such as
device parameters like (for example) Isrc:Mag.  These types of
parameters are the reason that I couldn't completely remove these
functions altogether.

But I did remove them from the deviceExpressionGroup, as the version of these
functions in that group were ONLY for global_params.  So, at this point the
deviceExpressionGroup::getGlobalParameterValue functions were dead code and
they had to go.