Restore "getImmutableValue" function calls
This commit is a fix for Issue #225:
"Fix the MEASURE/bad_dot_measure_line2 test"

In my various experimentations with global_params, at some point I
experimented with unconditionally elevating all .param in the top level
circuit context to .global_params.  I initially thought that this could
be a "quick and dirty" way of addressing Issue #167, which was:

"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

I quickly realized this approach was not the best way to go, for
multiple reasons.  One reason was that there are many contexts in xyce
where .param is allowed to be used, and .global_param is not, because
certain things are expected to be constant and from Xyce's point of
view, global_param = non-constant.

Before I gave up on the idea, however, I experimented with attempting to
change some parts of the code to allow non-constant parmaeters.  I did
this by replacing getImmutableValue function calls with getMutableValue
functions calls.  One of the places that I did this was in the .measure
code.  This mostly didn't cause any problems, but it did break the
MEASURE/bad_dot_measure_line2 test.

Earlier this week, I hacked that test so it wouldn't fail anymore, but
that was the wrong thing to do long term.  I did it at the time to make
testing stop complaining.  Unfortunately, that didn't quite work, as
this test is still complaining on some platforms.

This commit fixes it "the right way" by restoring all the
"getImmutableValue" calls.  Having them be "getMutableValue" didn't
serve any purpose, once I abandoned my first approach to Issue #167.  So
this commit puts it back the way it was.