Various efficiency improvements for expressions
This is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

That particular circuit was relatively small, 816 unknowns.  It also
consisted mostly of behavioral models.  (236 Bsrcs).    It is also a very long
running transient (1s), with a lot of behavior that forces small time steps.
So, under the best of circumstances it takes many hours to run.

The circuit failed to finish in 7.2, so it wasn't just slower.  But,
I was able to track down the convergence problems and fix them several months
ago, in Issue #108, which was entitled:
"Fix various problems in the limit operator in the new expression library"

Even after that fix however, Xyce was still running this circuit much more
slowly than 7.1.   I had some theories as to why this was, but didn't have
time to test them out until today.    Some of them proved correct, and so I've
managed to improve the situation.

This commit contains two things.

1.  another refactor of the "group" class, especially for the "device group".
This refactor eliminates 99% of the "find" commands being used by that class
during function evaluation.  To make this possible, a lot of the innards of
the "newExpression::getValuesFromGroup" function was moved into the group
classes, and this made that process much easier to optimize.  Also, I created
a new "setupGroup" function, that for some groups allows them to do this
optimization, and just do things like "find" during the setup.

2. The expression library, when called for derivatives, was allocating and
de-allocating the derivatives vector much more than it needed to.  It should
have only been allocating it once.   These allocations/deallocations were
happening b/c the Bsrc wants everything as a std::vector<double> and the
expression library is using std::vector<std::complex<double> > .  This
commit gets rid of most of those allocations/deallocations.  I didn't expect
this to matter much, but it showed up in the profiler.

Here are some results from running the customer circuit, but for just an early
part of the transient.    The original transient goes for 1s, but for testing
purposes, I shortened it to  5.0e-3s.  That shortened the runtime to minutes,
rather than hours, which was what I needed to quickly evaluate performance.
Here are runtime results:

Xyce 7.1:  119.961 seconds    <- old expression library
Xyce 7.2:  227.651 seconds    <- new expression library, with limit op bug
Xyce 7.3:  235.599 seconds    <- new expression lib, limit op bug fixed
Xyce dev (5/21/2021):  230.936 seconds  <- group refactor for global_param
Xyce dev (5/24/2021):  179.048 seconds  <- today's changes

The goal, of course is to get back to the performance of 7.1.  This commit
doesn't get Xyce there yet, but it is a step in the right direction.