Small memory optimization in "Depend" struct
This is related to various performance related issues that I am working
on.  This includes:

https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

The Depend struct is used by the device entity to help manage what it
calls "dependent parameters".  It stores metadata associated with
expressions, etc.

I noticed today that this structure contained a std::vector<double> and
a std::vector<std::string> objects, and these objects were no longer
needed post-new-expression-refactor.  The vector of doubles was for
values to be passed into the expression (which doesn't happen anymore)
and the vector of strings was of global_param names, which the
expression depended on.  For similar reasons, this wasn't needed.

But, despite not being needed, they were still sitting around and getting
allocated and set up.  This commit removes them.