Document legal usage of solution-dependent expressions
The reference guide section on expression usage in device parameters
has a list of contexts in which it is legal to use solution-dependent
expressions.

In answering a question on the google group, I notice that I had
failed to update this section correctly when I implemented the "charge
conservation" version of the capacitor for bug number 1112 in joseki's
bugzilla.  The list should include the capacitor's Q expression.

"Charge conservation" is a misnomer, because all versions of Xyce's
capacitor conserve charge, even when the capacitance is solution
dependent (I took great pains to assure that our capacitor conserves
charge in this case), but this one is HSPICE and LTSPICE compatible,
in that the user specifies an expression for the charge instead of the
capacitance.

Now the list of device parameters for which solution-dependent
expressions are legal contains this.

Note for future reference:  the solution-dependent capacitor's tests
are all in the "SOLN_DEP_CAP" directory of the test suite.  That
test includes a check that specifying a capacitor with charge Q
produces equivalent results to specifying a capacitor with a C
expression that is exactly the derivative with respect to voltage of
the other capacitor's Q.

https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1112