Remove a few valgrind errors
Two types of errors were removed in this commit.

(1) branches based on uninitialized values in the base expression
group.  This was in the new "putValues" function that I added to
this group yesterday.  I am not sure that I believed that these
valgrind errors were correct, but it was easy to make them go away.

(2) memory leak in the parser, associated with table files.  In
the parsing of tablefiles, I allocated a string that I neglected
to deallocate.   I am not sure when I introduced this, but my best
guess at the moment  is that I introdcued this problem in the
leadup to the 7.3 release.    Fixed now.