Fix small memory error.
The sdt12 test (both double and complex version) had a small mistake in it.

The small mistake had to do with the testing of the derivative calculation.
The derivative calculation (done by newExpression::evaluate) populates
a std::vector of derivative values.  In the sdt12 unit test, it was set up
to test a derivative array of size=2, but the expression only had a single
derivative.  So, its test for deriv[0] made sense and its test for deriv[1] did not.

This test did not throw a valgrind error, although arguably it should have.
The mistake was mostly harmless because the allocations for deriv were in
the unit test itself.  The expression library then resizes the deriv
array as needed.  So for this test, the expression library was internally
resizing the array from 2 to 1, and then only populating deriv[0].    I
guess b/c the resizing made the array smaller, rather than larger, and there
were no other memory allocations before the test went out of scope, prevented
the test of deriv[1] from throwing a valgrind error.  But it was still a
mistake, and fixing it was the right thing to do.  Fixed now.