Efficiency improvements for expression derivatives
This is related to this issue on the SRN gitlab:
https://gitlab.sandia.gov/Xyce/Xyce_SandiaModels/-/issues/5
"Xyce 7.2 is much slower than 7.1 on customer circuit"

The customer circuit associated with this issue has a large number of
Bsrcs, around 236, and it runs for a long transient.  As such, it
frequently calls the newExpression::evaluate function, which computes
both the value and the derivatives of the expression.    The vast
majority of expression use in Xyce is calls the
newExpression::evaluateFunction calll instead, which ONLY computes the
value.

In my first pass of the expression library, I didn't try very hard to
make the derivative calculation efficient, and I knew there were a few
things in the first cut that could be done more efficiently.  Today I
started working on one of those things, which was to change the code to
compute all the derivatives in a single AST traversal, rather than doing
a separate traversal for each derivative.     I also need to work on
doing the function evalation in the same traversal as the derivative
calculation, but that will come later.

This commit contains my first cut at doing this, multiple-derivative,
single traversal.  The old AST function (which did one at a time) was
called "dx()" and the new function is called "dx2()".   Unfortunately,
while this dx2() implementation passes all the unit
tests and regression tests, it is doing some unnecessary data
copies.  So, it isn't faster yet.  That will take another iteration or
two.

In the meantime, while working on this I observed several very obvious
places in the dx() functions where I could make things a bit faster.  In
a number of derivatives, I need to evaluate either the value or
derivative (val() or dx()) of one of the subordinate nodes, and then I
need to use them multiple times in the expanded derivative expression.
In a bunch of places, rather than call val() and dx() a single time, and
then use the result when needed, I was calling val() and dx() every
time.  So, a lot of wasted calls.  This commit gets rid of those, and it
made a real difference.

So, this commit is a checkpoint commit for the dx2() work, and a speedup
for the use of dx().  dx(), the old derivative function is still used
and the calls to dx2 are commented out.

With the above improvements, the runtime of the shortened transient is
another 20 seconds faster, roughly.  So, here are the current timings,
when running the initial  5.0e-3s of the circuit:

Xyce 7.1:  119.961 seconds    <- old expression library
Xyce 7.2:  227.651 seconds    <- new expression library, with limit op bug
Xyce 7.3:  235.599 seconds    <- new expression lib, limit op bug fixed
Xyce dev (5/21/2021):  230.936 seconds  <- group refactor for global_param
Xyce dev (5/24/2021):  179.048 seconds  <- yesterday's changes
Xyce dev (5/25/2021):  157.217 seconds  <- today's changes

As I stated in my previous commit message, the goal is to get back to
being at least as fast as 7.1.